organizing imports. Moving Gql resource closer to the location in which it is used. splitting out main source functionality from other, less interesting bits
