+++ pipeline/voyager_pipeline.py
@@ -3,38 +3,25 @@ from __future__ import annotations
 import logging
 import os
 
-import duckdb
-from voyager_client import CustomAuth
-
-logger = logging.getLogger(__name__)
-
-import json
-
 import dlt
-from analytics.saved_queries import (db_followed_companies,
-                                     get_dependency_from_db, log_current_jobs,
-                                     write_new_jobs_to_csv)
+import duckdb
+from client.helpers import LinkedInPaginator
+from client.voyager_client import CustomAuth
 from configuration.column_mapping import get_map_func
 from configuration.endpoint_conf import (data_selectors, dependencies,
                                          endpoints, mappings, total_paths)
 from configuration.pipeline_conf import API_BASE_URL, graphql_pagignator_config
+from data_utils import (db_followed_companies, get_dependency_from_db,
+                        log_current_jobs, write_new_jobs_to_csv)
 from dlt.sources.rest_api import rest_api_resources
 from dlt.sources.rest_api.typing import RESTAPIConfig
-from helpers import LinkedInPaginator
+from meta_resources import as_resource
+
+logger = logging.getLogger(__name__)
+
 
 # Resource Creation Functions
 # # These are config-driven resource generators
-
-
-def as_resource(name, data, **kwargs):
-    @dlt.resource(name=name, **kwargs)
-    def data_resource():
-        if isinstance(data, list):
-            yield data
-        else:
-            yield [data]
-
-    return data_resource
 
 
 def graphql_resource(source_name, inspect_response=False):
@@ -72,97 +59,6 @@ def graphql_resource(source_name, inspect_response=False):
     if include_from_parent:
         resource_config["include_from_parent"] = include_from_parent
     return resource_config
-
-
-@dlt.source
-def test_source(
-    source_name,
-    session,
-):
-    resources = [graphql_resource(source_name)]
-    config: RESTAPIConfig = {
-        "client": {
-            "base_url": f"{API_BASE_URL}",
-            "session": session,
-        },
-        # There doesn't seem to be a way to set
-        # write_disposition, merge strategy from the schema file
-        "resource_defaults": {
-            "write_disposition": {"disposition": "merge", "strategy": "scd2"}
-        },
-        "resources": resources,
-    }
-    resource_list = rest_api_resources(config)
-    return resource_list
-
-
-bucket_url = "file://Users/admin/Documents/csv_files"
-
-
-def file_resource(
-    file_name,
-):
-    with open(file_name) as f:
-        data = json.load(f)
-        recipes = []
-        mappings = []
-        all_fields_rows = []
-        for recipe_id, datum in data.items():
-            fields_details = datum["fields"]
-            fields = datum["fields"].keys()
-            recipe = [
-                {
-                    "table_name": "recipes",
-                    "recipe_id": recipe_id,
-                    "baseType": datum["baseType"],
-                    "fields": list(fields),
-                }
-            ]
-            fields_to_recipe = [
-                {
-                    "table_name": "fields_to_recipe",
-                    "recipe_id": recipe_id,
-                    "field": field,
-                }
-                for field in fields
-            ]
-            field_rows = []
-            for field, details in fields_details.items():
-                row = {"table_name": "fields", "field": field}
-                for name, value in details.items():
-                    row[name] = value
-                field_rows.append(row)
-            recipes.extend(recipe)
-            mappings.extend(fields_to_recipe)
-            all_fields_rows.extend(field_rows)
-
-        return recipes, mappings, all_fields_rows
-
-
-@dlt.source
-def schemata_resources():
-    import glob
-
-    files = glob.glob("data/endpoint_responses/schemata/*.json")
-    all_recipes = []
-    all_mappings = []
-    all_fields_rows = []
-    for file in files:
-        data = file_resource(file)
-        recipes, mappings, fields_rows = data
-        all_recipes.extend(recipes)
-        all_mappings.extend(mappings)
-        all_fields_rows.extend(fields_rows)
-
-    resources = []
-    kwargs = {"max_table_nesting": 0, "table_name": lambda row: row["table_name"]}
-    kwargs["primary_key"] = "recipe_id"
-    resources.append(as_resource("recipe_resource", all_recipes, **kwargs))
-    _ = kwargs.pop("primary_key")
-    resources.append(as_resource("mapping_resource", all_mappings, **kwargs))
-    kwargs["primary_key"] = "field"
-    resources.append(as_resource("field_resource", all_fields_rows, **kwargs))
-    return resources
 
 
 @dlt.source
